Consolidate two modules
