+++ agent/agent-profiler/agent-service-profiler/src/main/java/com/microsoft/applicationinsights/serviceprofilerapi/JfrProfilerServiceFactory.java
@@ -4,8 +4,7 @@
 package com.microsoft.applicationinsights.serviceprofilerapi;
 
 import com.azure.core.http.HttpPipeline;
-import com.microsoft.applicationinsights.profiler.ProfilerConfigurationHandler;
-import com.microsoft.applicationinsights.profiler.config.ServiceProfilerServiceConfig;
+import com.microsoft.applicationinsights.serviceprofilerapi.config.LocalConfig;
 import com.microsoft.applicationinsights.serviceprofilerapi.client.ServiceProfilerClient;
 import com.microsoft.applicationinsights.serviceprofilerapi.profiler.JfrProfiler;
 import com.microsoft.applicationinsights.serviceprofilerapi.upload.UploadService;
@@ -13,10 +12,6 @@ import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.function.Supplier;
 
-/**
- * Default ProfilerService factory loaded by a service loader, produces a Profiler Service based on
- * JFR.
- */
 public class JfrProfilerServiceFactory {
 
   @SuppressWarnings("TooManyParameters")
@@ -24,13 +19,14 @@ public class JfrProfilerServiceFactory {
       Supplier<String> appIdSupplier,
       ProfilerConfigurationHandler profilerConfigurationHandler,
       String processId,
-      ServiceProfilerServiceConfig config,
+      LocalConfig config,
       String machineName,
       String instrumentationKey,
       HttpPipeline httpPipeline,
       ScheduledExecutorService serviceProfilerExecutorService,
       String userAgent,
       String roleName) {
+
     ServiceProfilerClient serviceProfilerClient =
         new ServiceProfilerClient(
             config.getServiceProfilerFrontEndPoint(), instrumentationKey, httpPipeline, userAgent);
