+++ agent/agent-profiler/agent-service-profiler/src/main/java/com/microsoft/applicationinsights/serviceprofilerapi/ProfilerConfiguration.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.applicationinsights.profiler;
+package com.microsoft.applicationinsights.serviceprofilerapi;
 
 import com.squareup.moshi.Json;
 import java.util.Date;
