+++ agent/agent-profiler/agent-service-profiler/src/main/java/com/microsoft/applicationinsights/serviceprofilerapi/upload/ServiceProfilerIndex.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.applicationinsights.profiler.uploader;
+package com.microsoft.applicationinsights.serviceprofilerapi.upload;
 
 import java.util.HashMap;
 import java.util.Map;
