+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/profiler/GcEventMonitorTest.java
@@ -11,7 +11,7 @@ import com.microsoft.applicationinsights.alerting.AlertingSubsystem;
 import com.microsoft.applicationinsights.alerting.alert.AlertBreach;
 import com.microsoft.applicationinsights.alerting.analysis.TimeSource;
 import com.microsoft.applicationinsights.alerting.config.AlertingConfiguration;
-import com.microsoft.applicationinsights.profiler.config.AlertConfigParser;
+import com.microsoft.applicationinsights.serviceprofilerapi.config.AlertConfigParser;
 import com.microsoft.gcmonitor.GcCollectionEvent;
 import com.microsoft.gcmonitor.GcEventConsumer;
 import com.microsoft.gcmonitor.GcMonitorFactory;
