+++ settings.gradle.kts
@@ -49,7 +49,6 @@ include(":agent:agent-gc-monitor:gc-monitor-api")
 include(":agent:agent-gc-monitor:gc-monitor-core")
 include(":agent:agent-gc-monitor:gc-monitor-tests")
 
-include(":agent:agent-profiler:agent-profiler-api")
 include(":agent:agent-profiler:agent-alerting-api")
 include(":agent:agent-profiler:agent-diagnostics-api")
 include(":agent:agent-profiler:agent-alerting")
